+++ src/components/contexts/SInfoContext.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+
+const SInfoContext = React.createContext('SInfoContext');
+export default SInfoContext;